test(ROB-221): deduplicate market event helpers
