+++ tests/services/test_market_events_query_service.py
@@ -7,34 +7,13 @@ from decimal import Decimal
 
 import pytest
 import pytest_asyncio
-from sqlalchemy import delete, select
+
+from tests.market_events_test_helpers import clean_non_tradingview_market_events
 
 
 @pytest_asyncio.fixture(autouse=True)
 async def _clean_market_events(db_session):
-    from app.models.market_events import (
-        MarketEvent,
-        MarketEventIngestionPartition,
-        MarketEventValue,
-    )
-
-    non_tradingview_events = select(MarketEvent.id).where(
-        MarketEvent.source != "tradingview"
-    )
-    await db_session.execute(
-        delete(MarketEventValue).where(
-            MarketEventValue.event_id.in_(non_tradingview_events)
-        )
-    )
-    await db_session.execute(
-        delete(MarketEvent).where(MarketEvent.source != "tradingview")
-    )
-    await db_session.execute(
-        delete(MarketEventIngestionPartition).where(
-            MarketEventIngestionPartition.source != "tradingview"
-        )
-    )
-    await db_session.commit()
+    await clean_non_tradingview_market_events(db_session)
     yield
