+++ tests/test_market_events_router.py
@@ -2,66 +2,26 @@
 
 from __future__ import annotations
 
-from types import SimpleNamespace
-
 import pytest
 import pytest_asyncio
-from fastapi import FastAPI
 from fastapi.testclient import TestClient
-from sqlalchemy import delete, select
+
+from tests.market_events_test_helpers import (
+    build_market_events_app,
+    clean_non_tradingview_market_events,
+)
 
 
 @pytest_asyncio.fixture(autouse=True)
 async def _clean_market_events(db_session):
-    from app.models.market_events import (
-        MarketEvent,
-        MarketEventIngestionPartition,
-        MarketEventValue,
-    )
-
-    non_tradingview_events = select(MarketEvent.id).where(
-        MarketEvent.source != "tradingview"
-    )
-    await db_session.execute(
-        delete(MarketEventValue).where(
-            MarketEventValue.event_id.in_(non_tradingview_events)
-        )
-    )
-    await db_session.execute(
-        delete(MarketEvent).where(MarketEvent.source != "tradingview")
-    )
-    await db_session.execute(
-        delete(MarketEventIngestionPartition).where(
-            MarketEventIngestionPartition.source != "tradingview"
-        )
-    )
-    await db_session.commit()
+    await clean_non_tradingview_market_events(db_session)
     yield
-
-
-def _app() -> FastAPI:
-    from app.core.db import get_db
-    from app.routers import market_events
-    from app.routers.dependencies import get_authenticated_user
-
-    app = FastAPI()
-    app.include_router(market_events.router)
-    app.dependency_overrides[get_authenticated_user] = lambda: SimpleNamespace(id=7)
-
-    async def _override_get_db():
-        from app.core.db import AsyncSessionLocal
-
-        async with AsyncSessionLocal() as session:
-            yield session
-
-    app.dependency_overrides[get_db] = _override_get_db
-    return app
 
 
 @pytest.mark.integration
 def test_get_today_events_returns_empty_when_no_data(db_session):
     """Smoke test: route exists, returns empty events when DB has none."""
-    with TestClient(_app()) as client:
+    with TestClient(build_market_events_app()) as client:
         response = client.get(
             "/trading/api/market-events/today?on_date=2026-05-07",
         )
@@ -74,12 +34,7 @@ def test_get_today_events_returns_empty_when_no_data(db_session):
 @pytest.mark.integration
 def test_get_today_events_unauthorized_without_override():
     """Without dependency override, real auth dependency rejects unauthenticated request."""
-    from app.routers import market_events
-
-    app = FastAPI()
-    app.include_router(market_events.router)
-    # Do not override get_authenticated_user — it should reject calls without a token.
-    with TestClient(app) as client:
+    with TestClient(build_market_events_app(authenticated=False)) as client:
         response = client.get(
             "/trading/api/market-events/today?on_date=2026-05-07",
         )
@@ -88,7 +43,7 @@ def test_get_today_events_unauthorized_without_override():
 
 @pytest.mark.integration
 def test_get_range_events_validates_date_order(db_session):
-    with TestClient(_app()) as client:
+    with TestClient(build_market_events_app()) as client:
         response = client.get(
             "/trading/api/market-events/range?from_date=2026-05-08&to_date=2026-05-07",
         )
@@ -98,7 +53,7 @@ def test_get_range_events_validates_date_order(db_session):
 @pytest.mark.integration
 def test_get_today_events_filters_by_category_economic(db_session):
     """Smoke test: passing category=economic does not 400 and filters correctly."""
-    with TestClient(_app()) as client:
+    with TestClient(build_market_events_app()) as client:
         response = client.get(
             "/trading/api/market-events/today?on_date=2026-05-13&category=economic&market=global",
         )
@@ -110,7 +65,7 @@ def test_get_today_events_filters_by_category_economic(db_session):
 
 @pytest.mark.integration
 def test_get_today_events_rejects_unknown_category(db_session):
-    with TestClient(_app()) as client:
+    with TestClient(build_market_events_app()) as client:
         response = client.get(
             "/trading/api/market-events/today?on_date=2026-05-13&category=bogus",
         )
